+++ packages/ripple-binary-codec/HISTORY.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+* Support for the AMMClawback amendment (XLS-73).
 * Support for the `Batch` amendment (XLS-56).
 
 ## 2.2.0 (2024-12-23)